+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.58.4
+Add POST method to HttpMocker
+
+## 0.58.3
+fix declarative oauth initialization
+
 ## 0.58.2
 Integration tests: adding debug mode to improve logging